add cudnn convolution forward test cases
